requirements.txt
PyYAML==3.10
flickrapi==1.4.2
wsgiref==0.1.2
